chanve type of crc to allow -1 (Coverity)
git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/libgphoto2@15452 67ed7778-7388-44ab-90cf-0a291f65f57c

+++ camlibs/canon/serial.c
@@ -367,7 +367,7 @@ canon_serial_send_packet (Camera *camera, unsigned char type, unsigned char seq,
 			  unsigned char *pkt, int len)
 {
 	unsigned char *hdr = pkt - PKT_HDR_LEN;
-	unsigned short crc;
+	int crc;
 
 	hdr[PKT_TYPE] = type;
 	hdr[PKT_SEQ] = seq;
